[sculpt]
Introduce a latency mode that lights an LED upon each keydown. This
allows measurement, with a high-speed camera, of the delta between
QMK's detection of the keypress event and the USB stack and host
device.

I've measured about 5 ms between the keyboard's LED lighting and the
LED on a Raspberry Pi Zero with a simple program that reads events
directly from /dev/input/event0 and lights the Pi's onboard LED. Same
numbers on a Raspberry Pi 3B.